move overlap_area
