+++ data/rules.py
@@ -5,6 +5,7 @@ from numpy.random import randint
 from tqdm import trange
 from data.shapes import ShapeGenerator
 from data.relations import RelationGenerator
+from data.utils import overlap_area
 
 
 def generate_rules(data_args, rule_args) -> list[dict[str, list]]:
@@ -62,30 +63,12 @@ def generate_rules(data_args, rule_args) -> list[dict[str, list]]:
     return results
 
 
-def no_overlap(shapes, new_shape, exclude_shape: list = [], thres=0.2) -> bool:
+def no_overlap(shapes, new_shape, exclude_shape=None, thres=0.2) -> bool:
     if new_shape is None:
         return False
 
-    def overlap_area(bbox1, bbox2) -> float:
-        min_x1, max_y1 = bbox1[0]
-        max_x1, min_y1 = bbox1[1]
-        min_x2, max_y2 = bbox2[0]
-        max_x2, min_y2 = bbox2[1]
-
-        # Calculate the intersection coordinates
-        overlap_min_x = max(min_x1, min_x2)
-        overlap_max_x = min(max_x1, max_x2)
-        overlap_min_y = max(min_y1, min_y2)
-        overlap_max_y = min(max_y1, max_y2)
-
-        # Calculate the width and height of the overlap
-        overlap_width = max(0, overlap_max_x - overlap_min_x)
-        overlap_height = max(0, overlap_max_y - overlap_min_y)
-
-        # Calculate the area of the overlap
-        overlap_area = overlap_width * overlap_height
-
-        return overlap_area
+    if exclude_shape is None:
+        exclude_shape = []
 
     iou_sum = 0
     for cur_shape in shapes:
